Release 1.1.0
