+++ docs/locale/ja/LC_MESSAGES/history.po
@@ -10,7 +10,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: k2hr3client \n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-10-03 18:04+0900\n"
+"POT-Creation-Date: 2024-10-04 20:13+0900\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language: ja\n"
@@ -26,46 +26,62 @@ msgid "History"
 msgstr "歴史"
 
 #: ../../../HISTORY.rst:6
-msgid "1.0.3 (2024-10-03)"
+msgid "1.1.0 (2024-10-04)"
 msgstr ""
 
 #: ../../../HISTORY.rst:8
+msgid "Removes the deprecated Interfaces"
+msgstr "非推奨のインタフェースを削除"
+
+#: ../../../HISTORY.rst:9
+msgid "Removes the examples"
+msgstr "例示プログラムの削除"
+
+#: ../../../HISTORY.rst:10
+msgid "Updates the docs"
+msgstr "ドキュメントの更新"
+
+#: ../../../HISTORY.rst:13
+msgid "1.0.3 (2024-10-03)"
+msgstr ""
+
+#: ../../../HISTORY.rst:15
 msgid "Fixes the K2hr3 API version"
 msgstr "K2hr3 APIのバージョンを固定化"
 
-#: ../../../HISTORY.rst:9
+#: ../../../HISTORY.rst:16
 msgid "Deprecates K2hr3Resource.create_conf_resource params"
 msgstr "K2hr3Resource.create_conf_resourceのパラメータを一部非推奨化"
 
-#: ../../../HISTORY.rst:12
+#: ../../../HISTORY.rst:19
 msgid "1.0.2 (2024-09-26)"
 msgstr ""
 
-#: ../../../HISTORY.rst:14
+#: ../../../HISTORY.rst:21
 msgid "Fixes lint errors"
 msgstr "静的解析ツールのエラーを修正"
 
-#: ../../../HISTORY.rst:17
+#: ../../../HISTORY.rst:24
 msgid "1.0.1 (2024-09-25)"
 msgstr ""
 
-#: ../../../HISTORY.rst:19
+#: ../../../HISTORY.rst:26
 msgid "Fixes resource API bugs"
 msgstr "Resource APIの不具合を修正"
 
-#: ../../../HISTORY.rst:22
+#: ../../../HISTORY.rst:29
 msgid "1.0.0 (2024-08-28)"
 msgstr ""
 
-#: ../../../HISTORY.rst:24
+#: ../../../HISTORY.rst:31
 msgid "Supports the other APIs"
 msgstr "他のAPIをサポート"
 
-#: ../../../HISTORY.rst:27
+#: ../../../HISTORY.rst:34
 msgid "0.0.1 (2020-08-28)"
 msgstr ""
 
-#: ../../../HISTORY.rst:29
+#: ../../../HISTORY.rst:36
 msgid "Supports Token, Resource, Policy and Role API"
 msgstr "トークン、リソース、ポリシーおよびロールAPIをサポート"
