+++ src/k2hr3client/list.py
@@ -1,15 +1,15 @@
 # -*- coding: utf-8 -*-
 #
-# K2HDKC DBaaS based on Trove
+# k2hr3client - Python client for K2HR3 REST API
 #
 # Copyright 2020 Yahoo Japan Corporation
 # Copyright 2024 LY Corporation
 #
-# K2HDKC DBaaS is a Database as a Service compatible with Trove which
-# is DBaaS for OpenStack.
-# Using K2HR3 as backend and incorporating it into Trove to provide
-# DBaaS functionality. K2HDKC, K2HR3, CHMPX and K2HASH are components
-# provided as AntPickax.
+# K2HR3 is K2hdkc based Resource and Roles and policy Rules, gathers
+# common management information for the cloud.
+# K2HR3 can dynamically manage information as "who", "what", "operate".
+# These are stored as roles, resources, policies in K2hdkc, and the
+# client system can dynamically read and modify these information.
 #
 # For the full copyright and license information, please view
 # the license file that was distributed with this source code.
